Enum joytk::opts::SubCommand

pub enum SubCommand {
    Calibrate(Calibrate),
    Get,
    Set(Set),
    Monitor,
    PulseRate,
    Dump,
    Restore,
    Decode,
    Relay(Relay),
    Ringcon(Ringcon),
    Camera,
}

Variants

Calibrate(Calibrate)

Tuple Fields

0: Calibrate

Calibrate the controller

The calibration will be stored on the controller and used by the Switch.

Get

Print settings from the controller

Set(Set)

Tuple Fields

0: Set

Configure settings of the controller

Monitor

Show live inputs from the controller

PulseRate

Dump

Dump the memory of the controller to a binary file

Restore

Restore the memory of the controller from a dump file

Decode

Decode raw reports exchanged between the controller and the Switch

See the relay subcommand to record new traces.

See the trace/ folder for recorded dumps, and relay_joycon.py for capturing new dumps.

Relay(Relay)

Tuple Fields

0: Relay

Relay the bluetooth trafic between a controller and the Switch

Important commands are decoded and shown, and a full log can be recorded. See the decode subcommand to decode logs.

Ringcon(Ringcon)

Tuple Fields

0: Ringcon

Ringcon-specific actions

Camera